+++ server/src/main/java/br/com/lcano/centraldecontrole/dto/UsuarioConfigDTO.java
@@ -3,13 +3,15 @@ package br.com.lcano.centraldecontrole.dto;
 import br.com.lcano.centraldecontrole.domain.UsuarioConfig;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class UsuarioConfigDTO {
     private Long id;
 
     private int despesaNumeroMaxItemPagina;
 
-    private double despesaValorMetaMensal;
+    private BigDecimal despesaValorMetaMensal;
 
     private int despesaDiaPadraoVencimento;
